This is a pre-pre-pre alpha version of this model.  It will probably not be useful to anyone for anything just yet.
The 3d model is by Lee Elliot.  Everything else is just copied over from
the J3 cub for now.

The Rascal 110 is a 110" (2.8m) wingspan R/C airplane.  The goal of this
model is to demonstrate an R/C size airplane in FlightGear.  I (Curt) have
some flight data on this aircraft (and have access to fly one of these) so
I hope to build a dynamics model that closely resembles the real thing.

Part of the push for this model is to to be an example of using FlightGear for
UAV development, modeling, testing, training, and even in real world operation.